Support concatenated gizp and bzip2 input
Default configuration of apache-commons-compress doesn't support
concatenated gzip (or bzip2) files. The format can support to set
decompressConcatenated parameter while constructing instances.

In this fix, gzip and bzip2 files are created with
decompressConcatenated = true. There is no parameter for factory
method. So, it is required to set 'format' parameter to enable
this fix. Set gz or gzip to format parameter for gzip files and
set bz2 or bzip2 to format parameter for bzip2.