Add decompress_concatenated parameter
It is not user friendly to configure format parameter when
concatenated gzip and bzip2 format files are used.
So, add 'decompress_concatenated' parameter and the default
value of the parameter is true.

So, normally, 'format' parameter is not required when input
file can detect format automatically. And if first gzip/bzip2
entry is only required, then set the parameter to false
to disable this new feature.